Update flow diagram to support branching edges for conditionals
- If there is a conditional subflow, instead of creating nodes
for the start and end of the subflow, there will now be branching edges
for the condition nodes, representing the false and true cases.
- Adds an optional label attribute to the edges to support having
true and false labels for condition node edges.
- Modifies auto layout to reduce the amount of overlap that
occurs with the branching condition edges and labels.
- Also removed the startSubFlow check in the createNode method
since the start and end subflow nodes are always created by the
renderSubFlowNodes method.

Closes #28453

Signed-off-by: Greg Baroni <greg.baroni@appfolio.com>