ui: make Scope Details page's Description a textarea, like Roles'
The Scope Details page had its description field as a single line text
field. This is inconsistent with the Roles page, which has it as a
textarea allowing multiline editing.

It's true that the character limit is 255 (which is way too short and
should probably be changed imo) but newlines are useful in tweets too :)

Signed-off-by: Jade Lovelace <software@lfcode.ca>