Allow users to customize the footer of a login theme (#31390)
Previously, users had to copy the whole template.ftl file or resort to css/js hacks to
add a custom footer.

This PR adds new footer.ftl template to the base login theme to allow users
to configure a footer for the login box, as well as the page footer.

The new footer.ftl template contains a content macro that can be added at the bottom of the login page.

Adjusted template.ftl in base/login and keycloak.v2 theme to render footer.ftl macro.

Fixes #31390

Signed-off-by: Thomas Darimont <thomas.darimont@googlemail.com>