Add kc_action to redirect URI after a required action is cancelled (#31894)
Previously, the kc_action parameter was missing and
clients that wanted to know which action got cancelled needed to
remember which action was executed.

We now pass the last executed required action, which was cancelled
as the kc_action parameter in the redirect URI such that clients
can know about which action was cancelled by the user.

Fixes #31894

Signed-off-by: Thomas Darimont <thomas.darimont@googlemail.com>