Script Support for Content Security Policy Header
Extends the CSP builder to support script-src, extends the Security Headers Provider to permit the injection of script-src directives, and implements hash generation for the on-the-fly JavaScript used for browser history.

Adds script-src nonce support to Freemarker login templates and the OIDC redirect POST. Adds style-src nonce support to Freemarker login templates.

Closes #32123

Signed-off-by: James Shuriff <james@febyte.com>