Add support for RFC 8707 OAuth2 Resource Indicators (#14355)
- Adds ResourceIndicatorMapper OIDC protocol mapper
- Make AuthorizationEndpoint and Token Endpoint aware of resource parameter.
- Enhanced Authz request parsing to support repeated parameter (resource)

If the authorization request contains one or more resource parameters, those resource values are first checked against
the list of allowed resource indicator patterns for the given client.
If the resource indicators are legit for the client, the initially requested resource identifier set is stored in the client session.
Token Requests / Token Refresh Requests can later request access tokens minted to support resource indicators in the aud claim based on the resource indicators that were initially requested.

Note that this allows users to dynamically extend or narrow down
the audience based on the given resource identifiers in token / token refresh requests.

The allowed resource identifiers for a client are managed via the ResourceIndicatorMapper.
ResourceIndicatorMapper manages aud claim via resource parameters. Users can configure the list of allowed resource indicator patterns in the ResourceIndicatorMapper configuration for a client. Note that only resource indicators allowed by the mapper configuration can be requested for a client.

Fixes #14355

Signed-off-by: Thomas Darimont <thomas.darimont@googlemail.com>