Remove dependency to "DOM" lib in @keycloak/admin-client
Previously, to use @keycloak/admin-client with typescript, you needed to
use the "DOM" lib in your project to avoid type errors. This is not ideal
in a server-based project because by including the "DOM" lib, you also
enable all other Web APIs that are not available on the server, and
TypeScript will no longer report errors if they are used.

Now, @keycloak/admin-client only uses the "ESNext" lib. The other projects
still use the "DOM" lib as well.

Code changes were needed because for technical reasons, @types/node
does not include any of the type aliases.

Signed-off-by: Jan Melcher <jan.melcher@aeb.com>