Add client OIDC compatibility mode setting for configuring the header type in access tokens.
If this setting is off, the access token header type will be "at+jwt" in compliance with RFC 9068, see https://datatracker.ietf.org/doc/html/rfc9068#section-2.1. If the setting is on, the access token header type will be "JWT". The setting is on per default, meaning existing clients will issue access tokens with header type "JWT", whereas the setting is off for new clients such that the access token header type is "at+jwt".

Closes #36696

Signed-off-by: Laurids Møller Jepsen <laurids.jepsen@cryptomathic.com>