Fix account console for usage with secure-session client-policy (#37447)
Previously the missing state parameter caused issues when the client policy `secure-session` was enabled for the realm.

This adds a dummy state parameter for the initial server-side auth redirect when we check for an authenticated user in the account-console backend.

If the user is successfully authenticated the account-console backend will bootstrap the
actual account-console frontend, which will then
perform the actual auth code flow with a proper state handling.

Fixes #37447

Signed-off-by: Thomas Darimont <thomas.darimont@googlemail.com>