Enable audience aliases during signed jwt authentication
Due to changes in OAuth2 requirements during signed jwt authentication
(urn:ietf:params:oauth:client-assertion-type:jwt-bearer)
where the JWT token used can only have one item in the aud claim, we see
a need to allow the keycloak instance to accept multiple 'identities'.

The use case we are solving with this change is having multiple keycloak
instances externally synced running in different data centers.  They are
all behind a load balancer which accepts connections from
keycloak.env.example.com, and they reply with a token signed with their
hostname (keycloak.dc.env.example.com) .  They all have different
hostnames, and as such do not accept a common audi value.  We have
solved this in the past by having the clients include the names of all
the possible keycloak instances in the aud clame.  After the change in
OIDC 1.0 draft 36, this scheme would no longer work.

If this PR gets accepted, we could add keycloak.env.example.com to the
jwt_auth_additional_audiences realm attribute and have the clients use
that as the aud claim.  As far as I can see this would not affect the
security of the keycloak instance or the tokens generated.

Signed-off-by: Thorhallur Sverrisson <toti@toti.is>