Updated the WildFly subsystem module to avoid using deprecated functionality from wildfly-core that's been removed in the latest version
Replaced instances of the deprecated class ModuleIdentifier with strings.
For now, maintain the use of the deprecated ModuleDependency constructor (while now using the module identifier strings from above in place of the ModuleIdentifier instances) in order to maximise compatibility. This ensures support for both WildFly, up to at least 37, and EAP 8.
Moved all the module identifier strings to KeycloakDependencyProcessor so they're all in one place. Removed duplicates and update references from other classes.

Closes #41669

Signed-off-by: Peter Griffiths <peter.griffiths@1spatial.com>