Extend documentation on ServiceMonitor creation
The documentation so far only mentions how to disable the ServiceMonitor, leading to the assumption that it is enabled by default.

Getting a service monitor created also requires enabling metrics in keycloak, though:

https://github.com/keycloak/keycloak/blob/658faf210f2ff91ee254ff9fcc62ddfbcbcd9e42/operator/src/main/java/org/keycloak/operator/controllers/KeycloakServiceMonitorDependentResource.java#L50

While the missing setting creates a warning, the administrator should already have this information before applying the configuration.

Closes #45070

Signed-off-by: Jens Erat <email@jenserat.de>